Make main list styles easier to apply.
The css and classes for the main list have been adjusted.
They are now easier to apply to both the general main list
(which is on the list of tasklists page)
and the derivative main list
(which is on the list of tasks page).

In fact, because of the selectors I used, the main list
entry styles will apply to all child elements of a main
list element, even if they don't have the main list
entry class. (I used the asterisk selector for this.)
If the list entries need to look significantly different,
you would have to either override the styles I added,
or you would have to apply a class to the list element
that isn't "main-list".