Fix keys in the form response not matching WTForms forms.
A function and a dictionary were added so that the data in
ImmutableMultiDict from flask.request.form (which contains
the form inputs) are usable in the WTForms form
subclasses. This is meant to resolve the scenario where
the keys present in the flask.request.form don't match the
attributes in the WTForms form subclasses (which is the
case for TaskForm).