Configure cascade for tasklist model so tasks are deleted.
I configured the cascade ability for the tasklist model.
This is so that if a tasklist is deleted, all of its tasks
are also deleted. Unlike the previous configuration (see
the commit labeled "Prevent tasklists from being
unintentionally deleted."), the deletion of tasks should
not cause the deletion of their associated tasklists.