Implement better queries for db when searching.
Better methods of querying the database were implemented
for searching for tasklists and tasks. Prior to this, the
SQLAlemchemy "ilike" operator was being used for
searching. This caused problems when search terms contained
characters like "%" or "_", which functioned as wildcards
in the generated sql statement. Now, I am using the
"contains" operator with the autoescape ability turned on,
meaning that these characters will be treated as literals,
not wildcards. I also made sure that these queries are
case-insensitive when performed.