Add initial support for displaying a sorted list of tasks.
I added the initial support for displaying a list of tasks
that is sorted based on a particular attribute. The
tasklists table in the database now stores an integer
representing a method of sorting (whether by completion
date, id, etc.) The enum in queries.py associates the
integer values with the method of sorting.